Fixed warnings, rearranged some headers.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@21705 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/terminal/Coding.h
@@ -35,7 +35,7 @@
 #include <SupportDefs.h>
 #include <UTF8.h>
 
-#define M_UTF8 (uint32)(-1)
+#define M_UTF8 (-1)
 
 status_t	get_nth_encoding(int i, int *id);
